Rename posts file to post header
